+++ packages/camera/camera_avfoundation/test/avfoundation_camera_test.mocks.dart
@@ -351,4 +351,22 @@ class MockCameraApi extends _i1.Mock implements _i2.CameraApi {
             returnValueForMissingStub: _i4.Future<void>.value(),
           )
           as _i4.Future<void>);
+
+  @override
+  _i4.Future<bool> isZeroShutterLagSupported() =>
+      (super.noSuchMethod(
+            Invocation.method(#isZeroShutterLagSupported, []),
+            returnValue: _i4.Future<bool>.value(false),
+            returnValueForMissingStub: _i4.Future<bool>.value(false),
+          )
+          as _i4.Future<bool>);
+
+  @override
+  _i4.Future<void> setZeroShutterLagEnabled(bool? enabled) =>
+      (super.noSuchMethod(
+            Invocation.method(#setZeroShutterLagEnabled, [enabled]),
+            returnValue: _i4.Future<void>.value(),
+            returnValueForMissingStub: _i4.Future<void>.value(),
+          )
+          as _i4.Future<void>);
 }